refactor shared part
