+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/injection/StepQuizFeatureBuilder.kt
@@ -2,9 +2,7 @@ package org.hyperskill.app.step_quiz.injection
 
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
-import org.hyperskill.app.core.view.mapper.ResourceProvider
 import org.hyperskill.app.freemium.domain.interactor.FreemiumInteractor
-import org.hyperskill.app.notification.local.domain.interactor.NotificationInteractor
 import org.hyperskill.app.problems_limit.presentation.ProblemsLimitActionDispatcher
 import org.hyperskill.app.problems_limit.presentation.ProblemsLimitFeature
 import org.hyperskill.app.problems_limit.presentation.ProblemsLimitReducer
@@ -30,11 +28,9 @@ object StepQuizFeatureBuilder {
         stepQuizInteractor: StepQuizInteractor,
         stepQuizReplyValidator: StepQuizReplyValidator,
         currentProfileStateRepository: CurrentProfileStateRepository,
-        notificationInteractor: NotificationInteractor,
         freemiumInteractor: FreemiumInteractor,
         analyticInteractor: AnalyticInteractor,
         sentryInteractor: SentryInteractor,
-        resourceProvider: ResourceProvider
     ): Feature<StepQuizFeature.State, StepQuizFeature.Message, StepQuizFeature.Action> {
         val stepQuizReducer = StepQuizReducer(stepRoute, problemsLimitReducer)
         val stepQuizActionDispatcher = StepQuizActionDispatcher(
@@ -42,11 +38,9 @@ object StepQuizFeatureBuilder {
             stepQuizInteractor,
             stepQuizReplyValidator,
             currentProfileStateRepository,
-            notificationInteractor,
             freemiumInteractor,
             analyticInteractor,
-            sentryInteractor,
-            resourceProvider
+            sentryInteractor
         )
 
         return ReduxFeature(
