+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/presentation/StepQuizFeature.kt
@@ -6,7 +6,6 @@ import org.hyperskill.app.step.domain.model.Step
 import org.hyperskill.app.step.domain.model.StepContext
 import org.hyperskill.app.step.domain.model.StepRoute
 import org.hyperskill.app.step_quiz.domain.model.attempts.Attempt
-import org.hyperskill.app.step_quiz.domain.model.permissions.StepQuizUserPermissionRequest
 import org.hyperskill.app.step_quiz.domain.model.submissions.Reply
 import org.hyperskill.app.step_quiz.domain.model.submissions.Submission
 import org.hyperskill.app.step_quiz.domain.validation.ReplyValidationResult
@@ -77,21 +76,9 @@ interface StepQuizFeature {
         data class SyncReply(val reply: Reply) : Message
 
         /**
-         * Request user permission
+         * Reset code
          */
-        data class RequestUserPermission(val userPermissionRequest: StepQuizUserPermissionRequest) : Message
-        data class RequestUserPermissionResult(
-            val userPermissionRequest: StepQuizUserPermissionRequest,
-            val isGranted: Boolean
-        ) : Message
-
-        /**
-         * Show problem of day solve modal
-         */
-
-        data class ShowProblemOfDaySolvedModal(val earnedGemsText: String) : Message
-
-        object ProblemOfDaySolvedModalGoBackClicked : Message
+        data class RequestResetCodeResult(val isGranted: Boolean) : Message
 
         /**
          * Daily limit reached modal
@@ -110,8 +97,6 @@ interface StepQuizFeature {
          */
         object ClickedCodeDetailsEventMessage : Message
         object ClickedRetryEventMessage : Message
-        object DailyStepCompletedModalShownEventMessage : Message
-        object DailyStepCompletedModalHiddenEventMessage : Message
         object ProblemsLimitReachedModalShownEventMessage : Message
         object ProblemsLimitReachedModalHiddenEventMessage : Message
 
@@ -140,11 +125,6 @@ interface StepQuizFeature {
             val submission: Submission
         ) : Action
 
-        data class RequestUserPermissionResult(
-            val userPermissionRequest: StepQuizUserPermissionRequest,
-            val isGranted: Boolean
-        ) : Action
-
         /**
          * Analytic
          */
@@ -158,9 +138,7 @@ interface StepQuizFeature {
         sealed interface ViewAction : Action {
             object ShowNetworkError : ViewAction // error
 
-            data class RequestUserPermission(val userPermissionRequest: StepQuizUserPermissionRequest) : ViewAction
-
-            data class ShowProblemOfDaySolvedModal(val earnedGemsText: String) : ViewAction
+            object RequestResetCode : ViewAction
 
             object ShowProblemsLimitReachedModal : ViewAction
 
@@ -169,8 +147,6 @@ interface StepQuizFeature {
             ) : ViewAction
 
             sealed interface NavigateTo : ViewAction {
-                object Back : NavigateTo
-
                 object Home : NavigateTo
 
                 data class StepScreen(val stepRoute: StepRoute) : NavigateTo
